packages/backend/pyproject.toml
[project]
name = "sayai-api"
version = "0.1.0"
description = "SayAi AI-OS API — Phases 1–3 (foundation + runtime + RAG/workflows)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1",
    "uvicorn[standard]>=0.32.0,<1",
    "sqlalchemy>=2.0.36,<3",
    "psycopg[binary]>=3.2.3,<4",
    "alembic>=1.14.0,<2",
    "pydantic-settings>=2.6.0,<3",
    "python-jose[cryptography]>=3.3.0,<4",
    "passlib[bcrypt]>=1.7.4,<2",
    "redis>=5.2.0,<6",
    "httpx>=0.28.0,<1",
    "email-validator>=2.2.0,<3",
    "litellm>=1.55.0,<2",
    "PyYAML>=6.0.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0,<9",
    "pytest-asyncio>=0.24.0,<1",
    "httpx>=0.28.0,<1",
    "ruff>=0.8.0,<1",
]

[build-system]
requires = ["setuptools>=75.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
